Modify web app folder structure - create core subfolder - move Root under web app - create msf-aggregate-data subfolder
