+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2020-12-01T09:53:24.942Z\n"
-"PO-Revision-Date: 2020-12-01T09:53:24.942Z\n"
+"POT-Creation-Date: 2020-12-02T16:23:09.972Z\n"
+"PO-Revision-Date: 2020-12-02T16:23:09.972Z\n"
 
 msgid "Field {{field}} cannot be blank"
 msgstr ""
@@ -1370,6 +1370,9 @@ msgstr ""
 msgid ""
 "Define who are the custodians of this instance metadata. Custodians approve "
 "metadata requests coming from other DHIS instances."
+msgstr ""
+
+msgid "Dashboard"
 msgstr ""
 
 msgid "Please fix the issues before testing the connection"
