+++ i18n/fr.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: i18next-conv\n"
-"POT-Creation-Date: 2020-12-01T09:53:24.942Z\n"
+"POT-Creation-Date: 2020-12-02T16:09:04.218Z\n"
 "PO-Revision-Date: 2020-07-10T06:53:30.625Z\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -1373,6 +1373,9 @@ msgstr ""
 msgid ""
 "Define who are the custodians of this instance metadata. Custodians approve "
 "metadata requests coming from other DHIS instances."
+msgstr ""
+
+msgid "Dashboard"
 msgstr ""
 
 msgid "Please fix the issues before testing the connection"
