+++ src/presentation/webapp/msf-aggregate-data/pages/MSFHomePage.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const MSFHomePage: React.FC = () => {
+    return <div>MSF!!!</div>;
+};
